Added conditional codes and check flags from CMP
Added conditional codes and check the flags from the compare operation
TODO: link the mux call to the conditional using cond_result to ensure that the condition is true for execution of the mux otherwise it skips and increments program counter. but we also need to consider that the mux should execute at least once to ensure that any non conditioned instructions can get through at the beginning. maybe modify the value of cond_result or set it as 1 initially